ml-version/requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0

# Deep Learning
torch>=2.0.0
tensorflow>=2.13.0
keras>=2.13.0

# Reinforcement Learning
gymnasium>=0.28.0
stable-baselines3>=2.0.0

# Data processing
scikit-learn>=1.3.0
ta>=0.11.0  # Technical Analysis Library

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
tensorboard>=2.13.0

# News data collection
newsapi-python>=0.2.7
gnews>=0.3.7
beautifulsoup4>=4.12.0
lxml>=4.9.0

# NLP for sentiment analysis
transformers>=4.30.0
sentencepiece>=0.1.99

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
PyYAML>=6.0

# Testing (optional)
pytest>=7.4.0
